+++ plugins/source/aws/client/services.go
@@ -15,7 +15,9 @@ import (
 	"github.com/aws/aws-sdk-go-v2/service/appsync"
 	"github.com/aws/aws-sdk-go-v2/service/athena"
 	"github.com/aws/aws-sdk-go-v2/service/autoscaling"
+	"github.com/aws/aws-sdk-go-v2/service/autoscalingplans"
 	"github.com/aws/aws-sdk-go-v2/service/backup"
+	"github.com/aws/aws-sdk-go-v2/service/batch"
 	"github.com/aws/aws-sdk-go-v2/service/cloudformation"
 	"github.com/aws/aws-sdk-go-v2/service/cloudfront"
 	"github.com/aws/aws-sdk-go-v2/service/cloudhsmv2"
@@ -123,6 +125,8 @@ func initServices(region string, c aws.Config) Services {
 		Appsync:                   appsync.NewFromConfig(awsCfg),
 		Athena:                    athena.NewFromConfig(awsCfg),
 		Autoscaling:               autoscaling.NewFromConfig(awsCfg),
+		Autoscalingplans:          autoscalingplans.NewFromConfig(awsCfg),
+		Batch:                     batch.NewFromConfig(awsCfg),
 		Backup:                    backup.NewFromConfig(awsCfg),
 		Cloudformation:            cloudformation.NewFromConfig(awsCfg),
 		Cloudfront:                cloudfront.NewFromConfig(awsCfg),
@@ -228,7 +232,9 @@ type Services struct {
 	Appsync                   services.AppsyncClient
 	Athena                    services.AthenaClient
 	Autoscaling               services.AutoscalingClient
+	Autoscalingplans          services.AutoscalingplansClient
 	Backup                    services.BackupClient
+	Batch                     services.BatchClient
 	Cloudformation            services.CloudformationClient
 	Cloudfront                services.CloudfrontClient
 	Cloudhsmv2                services.Cloudhsmv2Client