+++ plugins/source/aws/resources/services/batch/job_definitions.go
@@ -1,0 +1,72 @@
+package batch
+
+import (
+	"context"
+
+	"github.com/aws/aws-sdk-go-v2/aws"
+	"github.com/aws/aws-sdk-go-v2/service/batch"
+	"github.com/aws/aws-sdk-go-v2/service/batch/types"
+	"github.com/cloudquery/cloudquery/plugins/source/aws/client"
+	"github.com/cloudquery/plugin-sdk/schema"
+	"github.com/cloudquery/plugin-sdk/transformers"
+)
+
+func JobDefinitions() *schema.Table {
+	tableName := "aws_batch_job_definitions"
+	return &schema.Table{
+		Name:        tableName,
+		Description: `https://docs.aws.amazon.com/batch/latest/APIReference/API_DescribeJobDefinitions.html`,
+		Resolver:    fetchBatchJobDefinitions,
+		Multiplex:   client.ServiceAccountRegionMultiplexer(tableName, "batch"),
+		Transform:   transformers.TransformWithStruct(&types.JobDefinition{}),
+		Columns: []schema.Column{
+			client.DefaultAccountIDColumn(false),
+			client.DefaultRegionColumn(false),
+			{
+				Name:     "tags",
+				Type:     schema.TypeJSON,
+				Resolver: resolveBatchJobDefinitionTags,
+			},
+			{
+				Name:     "arn",
+				Type:     schema.TypeString,
+				Resolver: schema.PathResolver("JobDefinitionArn"),
+				CreationOptions: schema.ColumnCreationOptions{
+					PrimaryKey: true,
+				},
+			},
+		},
+	}
+}
+
+func fetchBatchJobDefinitions(ctx context.Context, meta schema.ClientMeta, parent *schema.Resource, res chan<- any) error {
+	config := batch.DescribeJobDefinitionsInput{
+		MaxResults: aws.Int32(100),
+	}
+	c := meta.(*client.Client)
+	svc := c.Services().Batch
+	p := batch.NewDescribeJobDefinitionsPaginator(svc, &config)
+	for p.HasMorePages() {
+		response, err := p.NextPage(ctx)
+		if err != nil {
+			return err
+		}
+		res <- response.JobDefinitions
+	}
+	return nil
+}
+
+func resolveBatchJobDefinitionTags(ctx context.Context, meta schema.ClientMeta, resource *schema.Resource, c schema.Column) error {
+	cl := meta.(*client.Client)
+	svc := cl.Services().Batch
+	summary := resource.Item.(types.JobDefinition)
+
+	input := batch.ListTagsForResourceInput{
+		ResourceArn: summary.JobDefinitionArn,
+	}
+	output, err := svc.ListTagsForResource(ctx, &input)
+	if err != nil {
+		return err
+	}
+	return resource.Set(c.Name, output.Tags)
+}