+++ plugins/source/facebookmarketing/rest/ad_study.go
@@ -57,7 +57,7 @@ func (facebookClient *FacebookClient) ListAdStudies(ctx context.Context, page st
 		RawQuery: query.Encode(),
 	}
 
-	request, err := http.NewRequestWithContext(ctx, http.MethodGet, u.String() /*body*/, nil)
+	request, err := http.NewRequestWithContext(ctx, http.MethodGet, u.String() /* body */, nil)
 	if err != nil {
 		return nil, "", sanitizeUrlError(err)
 	}