+++ plugins/source/facebookmarketing/rest/promote_page.go
@@ -186,7 +186,7 @@ func (facebookClient *FacebookClient) ListPromotePages(ctx context.Context, page
 		RawQuery: query.Encode(),
 	}
 
-	request, err := http.NewRequestWithContext(ctx, http.MethodGet, u.String() /*body*/, nil)
+	request, err := http.NewRequestWithContext(ctx, http.MethodGet, u.String() /* body */, nil)
 	if err != nil {
 		return nil, "", sanitizeUrlError(err)
 	}