+++ plugins/source/facebookmarketing/rest/publisher_block_list.go
@@ -49,7 +49,7 @@ func (facebookClient *FacebookClient) ListPublisherBlockLists(ctx context.Contex
 		RawQuery: query.Encode(),
 	}
 
-	request, err := http.NewRequestWithContext(ctx, http.MethodGet, u.String() /*body*/, nil)
+	request, err := http.NewRequestWithContext(ctx, http.MethodGet, u.String() /* body */, nil)
 	if err != nil {
 		return nil, "", sanitizeUrlError(err)
 	}