+++ plugins/source/shopify/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [1.1.3](https://github.com/cloudquery/cloudquery/compare/plugins-source-shopify-v1.1.2...plugins-source-shopify-v1.1.3) (2023-03-23)
+
+
+### Bug Fixes
+
+* **shopify:** Pagination should not include params in following pages ([#9389](https://github.com/cloudquery/cloudquery/issues/9389)) ([45e0c54](https://github.com/cloudquery/cloudquery/commit/45e0c5434503c2945921e6cc2b76b79a6bae6e2d)), closes [#9373](https://github.com/cloudquery/cloudquery/issues/9373)
+
+## [1.1.2](https://github.com/cloudquery/cloudquery/compare/plugins-source-shopify-v1.1.1...plugins-source-shopify-v1.1.2) (2023-03-21)
+
+
+### Bug Fixes
+
+* **deps:** Update module github.com/cloudquery/plugin-sdk to v1.44.0 ([#9167](https://github.com/cloudquery/cloudquery/issues/9167)) ([49d6477](https://github.com/cloudquery/cloudquery/commit/49d647730a85ea6fae51e97194ba61c0625d1331))
+
 ## [1.1.1](https://github.com/cloudquery/cloudquery/compare/plugins-source-shopify-v1.1.0...plugins-source-shopify-v1.1.1) (2023-03-14)